requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
matplotlib>=3.7.0

# Deep Learning
torch>=2.0.0
tensorboard>=2.12.0

# Optimization
osqp>=0.6.3

# Robotics
mujoco>=2.3.0

# Pinocchio (robot dynamics) - install via conda
# conda install -c conda-forge pinocchio

# Isaac Gym (for RL training) - install from NVIDIA
# Download from: https://developer.nvidia.com/isaac-gym

# Optional: for visualization
# open3d>=0.17.0
# pyvista>=0.38.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
mypy>=1.0.0
